refactor: extract scroll spy logic into reusable hook
