+++ src/app/como-eu-escolhi-lidar-com-pessoas/page.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState, type ComponentProps } from "react";
 
 import { AsideNavigationMenu } from "../(shared)/components/ui/aside-navigation-menu";
+import { useScrollSpy } from "@/hooks/article-progress";
 
 export default async function Home() {
   const navigationItems: ComponentProps<
@@ -20,28 +21,8 @@ export default async function Home() {
     },
   ];
 
-  const [activeItemIndex, setActiveItemIndex] = useState(0);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const scrollPosition = window.scrollY;
-      const sections = document.querySelectorAll("h4[id]");
-
-      let newActiveIndex = 0;
-      sections.forEach((section, index) => {
-        const sectionTop = section.getBoundingClientRect().top + scrollPosition;
-        if (sectionTop <= scrollPosition + window.innerHeight / 2) {
-          newActiveIndex = index;
-        }
-      });
-
-      setActiveItemIndex(newActiveIndex);
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
-
+  const { activeItemIndex } = useScrollSpy()
+  
   return (
     <>
       <main className="flex col-span-12 sm:col-span-9 py-8">
