+++ src/app/type-narrowing-vs-type-widening/page.tsx
@@ -9,6 +9,7 @@ import LetExample from "./mdx/let-example.mdx";
 import LetExampleFix from "./mdx/let-example-fix.mdx";
 import { ExternalRedirectIcon } from "@/app/(shared)/components/icons/external-redirect";
 import { AsideNavigationMenu } from "../(shared)/components/ui/aside-navigation-menu";
+import { useScrollSpy } from "@/hooks/article-progress";
 
 export default async function Home() {
   const navigationItems: ComponentProps<
@@ -27,27 +28,7 @@ export default async function Home() {
       withExternalRedirect: true,
     },
   ];
-  const [activeItemIndex, setActiveItemIndex] = useState(0);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const scrollPosition = window.scrollY;
-      const sections = document.querySelectorAll("h4[id]");
-
-      let newActiveIndex = 0;
-      sections.forEach((section, index) => {
-        const sectionTop = section.getBoundingClientRect().top + scrollPosition;
-        if (sectionTop <= scrollPosition + window.innerHeight / 2) {
-          newActiveIndex = index;
-        }
-      });
-
-      setActiveItemIndex(newActiveIndex);
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
+  const { activeItemIndex } = useScrollSpy()
 
   return (
     <>
