+++ src/app/type-vs-interface/page.tsx
@@ -32,6 +32,7 @@ import FunctionsWithInterface from "./mdx/functions-with-interface.mdx";
 
 import { AsideNavigationMenu } from "../(shared)/components/ui/aside-navigation-menu";
 import { useState } from "react";
+import { useScrollSpy } from "@/hooks/article-progress";
 
 export default function Home() {
   const navigationItems = [
@@ -86,27 +87,7 @@ export default function Home() {
     typeof AsideNavigationMenu
   >["navigationItems"];
 
-  const [activeItemIndex, setActiveItemIndex] = useState(0);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const scrollPosition = window.scrollY;
-      const sections = document.querySelectorAll("h4[id]");
-
-      let newActiveIndex = 0;
-      sections.forEach((section, index) => {
-        const sectionTop = section.getBoundingClientRect().top + scrollPosition;
-        if (sectionTop <= scrollPosition + window.innerHeight / 2) {
-          newActiveIndex = index;
-        }
-      });
-
-      setActiveItemIndex(newActiveIndex);
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
+ const { activeItemIndex } = useScrollSpy()
 
 
   return (
